+++ src/lib/query/proposal.ts
@@ -21,6 +21,7 @@ export const getRelatedProposalsByContractAddressPagination = graphql(`
         account {
           address
         }
+        is_expedited
       }
       proposal_id
       resolved_height
@@ -58,12 +59,8 @@ export const getProposalsByWalletAddressPagination = graphql(`
       deposit_end_time
       type
       id
-      contract_proposals {
-        resolved_height
-      }
-      code_proposals {
-        resolved_height
-      }
+      is_expedited
+      resolved_height
     }
   }
 `);
@@ -76,6 +73,51 @@ export const getProposalsCountByWalletAddress = graphql(`
       aggregate {
         count
       }
+    }
+  }
+`);
+
+export const getProposalList = graphql(`
+  query getProposalList(
+    $expression: proposals_bool_exp
+    $offset: Int!
+    $pageSize: Int!
+  ) {
+    proposals(
+      where: $expression
+      order_by: { id: desc }
+      offset: $offset
+      limit: $pageSize
+    ) {
+      type
+      id
+      title
+      voting_end_time
+      deposit_end_time
+      resolved_height
+      status
+      is_expedited
+      account {
+        address
+      }
+    }
+  }
+`);
+
+export const getProposalListCount = graphql(`
+  query getProposalListCount($expression: proposals_bool_exp) {
+    proposals_aggregate(where: $expression) {
+      aggregate {
+        count
+      }
+    }
+  }
+`);
+
+export const getProposalTypes = graphql(`
+  query getProposalTypes {
+    proposals(distinct_on: type) {
+      type
     }
   }
 `);