+++ src/lib/utils/index.ts
@@ -27,3 +27,4 @@ export * from "./codePermission";
 export * from "./executeTags";
 export * from "./cw2";
 export * from "./bech32";
+export * from "./proposal";